More inputmessagecontents and also inlinekeyboard* to single files
